refactoring the basic logic to a helper class
